ListBuckets: Update the function to filter by retion
AWS does not filter the list of Buckets by region so we have to do it manually
